streamlit/requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
google-cloud-bigquery>=3.11.0
google-auth>=2.0.0
